+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # DEV
 # 0.0.6
+- Flatten settings evaluation
 - Move module to top-level package
 # 0.0.5
 - Add psycopg2 as default db backend